+++ gdal/frmts/mrf/marfa_dataset.cpp
@@ -1639,12 +1639,12 @@ CPLErr GDALMRFDataset::AddVersion()
     // Hides the dataset variables with the same name
     VSILFILE *l_ifp = IdxFP();
 
-    void *tbuff = CPLMalloc(idxSize);
+    void *tbuff = CPLMalloc(static_cast<size_t>(idxSize));
     VSIFSeekL(l_ifp, 0, SEEK_SET);
-    VSIFReadL(tbuff, 1, idxSize, l_ifp);
+    VSIFReadL(tbuff, 1, static_cast<size_t>(idxSize), l_ifp);
     verCount++; // The one we write
     VSIFSeekL(l_ifp, idxSize * verCount, SEEK_SET); // At the end, this can mess things up royally
-    VSIFWriteL(tbuff, 1, idxSize, l_ifp);
+    VSIFWriteL(tbuff, 1, static_cast<size_t>(idxSize), l_ifp);
     CPLFree(tbuff);
     return CE_None;
 }
@@ -1697,12 +1697,12 @@ CPLErr GDALMRFDataset::WriteTile(void *buff, GUIntBig infooffset, GUIntBig size)
 	// tinfo contains the current info or 0,0
 	if (tinfo.size == GIntBig(net64(size))) { // Might be the same, read and compare
 	    if (size != 0) {
-		tbuff = CPLMalloc(size);
+		tbuff = CPLMalloc(static_cast<size_t>(size));
 		// Use the temporary buffer, we can't have a versioned cache !!
 		VSIFSeekL(l_dfp, infooffset, SEEK_SET);
-		VSIFReadL(tbuff, 1, size, l_dfp);
+		VSIFReadL(tbuff, 1, static_cast<size_t>(size), l_dfp);
 		// Need to write it if not the same
-		new_tile = (0 != memcmp(buff, tbuff, size));
+		new_tile = (0 != memcmp(buff, tbuff, static_cast<size_t>(size)));
 		CPLFree(tbuff);
 	    }
 	    else {
@@ -1732,7 +1732,7 @@ CPLErr GDALMRFDataset::WriteTile(void *buff, GUIntBig infooffset, GUIntBig size)
 	// Theese statements are the critical MP section for the data file
 	VSIFSeekL(l_dfp, 0, SEEK_END);
 	GUIntBig offset = VSIFTellL(l_dfp);
-	if (size != VSIFWriteL(buff, 1, size, l_dfp))
+	if (static_cast<size_t>(size) != VSIFWriteL(buff, 1, static_cast<size_t>(size), l_dfp))
 	    ret = CE_Failure;
 
 	tinfo.offset = net64(offset);
@@ -1743,11 +1743,11 @@ CPLErr GDALMRFDataset::WriteTile(void *buff, GUIntBig infooffset, GUIntBig size)
 	if (mp_safe) {
 	    // Allocate the temp buffer if we haven't done so already
 	    if (!tbuff)
-		tbuff = CPLMalloc(size);
+		tbuff = CPLMalloc(static_cast<size_t>(size));
 	    VSIFSeekL(l_dfp, offset, SEEK_SET);
-	    VSIFReadL(tbuff, 1, size, l_dfp);
+	    VSIFReadL(tbuff, 1, static_cast<size_t>(size), l_dfp);
 	    // If memcmp returns zero, verify passed
-	    if (!memcmp(buff, tbuff, size)) {
+	    if (!memcmp(buff, tbuff, static_cast<size_t>(size))) {
 		CPLFree(tbuff);
 		tbuff = NULL; // Triggers the loop termination
 	    }
@@ -1842,7 +1842,7 @@ CPLErr GDALMRFDataset::ReadTileIdx(ILIdx &tinfo, const ILSize &pos, const ILImag
     offset = (offset / CPYSZ) * CPYSZ;
     GIntBig size = MIN(size_t(CPYSZ), size_t(bias - offset));
     size /= sizeof(ILIdx); // In records
-    vector<ILIdx> buf(size);
+    vector<ILIdx> buf(static_cast<size_t>(size));
     ILIdx *buffer = &buf[0]; // Buffer to copy the source to the clone index
 
 
@@ -1856,7 +1856,7 @@ CPLErr GDALMRFDataset::ReadTileIdx(ILIdx &tinfo, const ILSize &pos, const ILImag
     }
 
     VSIFSeekL(srcidx, offset, SEEK_SET);
-    size = VSIFReadL(buffer, sizeof(ILIdx), size, srcidx);
+    size = VSIFReadL(buffer, sizeof(ILIdx), static_cast<size_t>(size), srcidx);
     if (size != GIntBig(buf.size())) {
 	CPLError(CE_Failure, CPLE_FileIO, "Can't read cloned source index");
 	return CE_Failure; // Source reported the error
@@ -1870,7 +1870,7 @@ CPLErr GDALMRFDataset::ReadTileIdx(ILIdx &tinfo, const ILSize &pos, const ILImag
 
     // Write it in the right place in the local index file
     VSIFSeekL(l_ifp, bias + offset, SEEK_SET);
-    size = VSIFWriteL(&buf[0], sizeof(ILIdx), size, l_ifp);
+    size = VSIFWriteL(&buf[0], sizeof(ILIdx), static_cast<size_t>(size), l_ifp);
     if (size != GIntBig(buf.size())) {
 	CPLError(CE_Failure, CPLE_FileIO, "Can't write to cloning MRF index");
 	return CE_Failure; // Source reported the error
